Updated for GGS 1.0.0b2 + /stats now shows your rank + GameMode uses getMain()

+++ src/com/gamezgalaxy/ctf/gamemode/ctf/CTF.java
@@ -112,7 +112,7 @@ public class CTF extends Gamemode {
 		} catch (IOException e1) {
 			e1.printStackTrace();
 		}
-		goal = mingoal + (int)(RANDOM.nextInt() * ((maxgoal - mingoal) + 1));
+		goal = mingoal + (int)(RANDOM.nextDouble() * ((maxgoal - mingoal) + 1));
 		running = true;
 		Thread run = new Checker();
 		run.start();
